[update] move Logger interface from usecases to interfaces
