+++ app/interfaces/logger.go
@@ -1,4 +1,4 @@
-package usecases
+package interfaces
 
 // A Logger belong to the usecases layer.
 type Logger interface {
